backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
torch==2.10.0
torchvision==0.25.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
scikit-image==0.22.0
Pillow==10.1.0

# 3D Processing
open3d==0.18.0
trimesh==4.0.5
pymeshlab==2023.12

# Computer Vision
mediapipe==0.10.8
dlib==19.24.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Task Queue
celery==5.3.4
flower==2.0.1

# Cloud Storage
boto3==1.29.7
azure-storage-blob==12.19.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==25.1.0
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
